Reorganize entity manager folder
