+++ discovery-provider/src/tasks/entity_manager/entities/developer_app.py
