+++ discovery-provider/src/tasks/entity_manager/entities/user.py
